+++ packages/gestalt/src/Icon.tsx
@@ -1,8 +1,5 @@
-import classnames from 'classnames';
-import styles from './Icon.css';
 import icons from './icons/index';
-import vrIcons from './icons-vr-theme/index';
-import useInExperiment from './useInExperiment';
+import InternalIcon from './sharedSubcomponents/InternalIcon';
 
 export type IconColor =
   | 'default'
@@ -60,32 +57,6 @@ type Props = {
 // @ts-expect-error - TS2322 - Type 'string[]' is not assignable to type 'readonly ("replace" | "search" | "link" | "text" | "dash" | "3D" | "3D-move" | "360" | "accessibility" | "ad" | "ad-group" | "add" | "add-circle" | "add-layout" | "add-pin" | "add-section" | ... 317 more ... | "wave")[]'.
 const IconNames: ReadonlyArray<keyof typeof icons> = Object.keys(icons);
 
-const flipOnRtlIconNames = [
-  'ads-stats',
-  'ads-overview',
-  'arrow-back',
-  'arrow-circle-forward',
-  'arrow-end',
-  'arrow-forward',
-  'arrow-start',
-  'arrow-up-right',
-  'compose',
-  'directional-arrow-left',
-  'directional-arrow-right',
-  'flipVertical',
-  'hand-pointing',
-  'link',
-  'mute',
-  'reorder-images',
-  'send',
-  'sound',
-  'speech',
-  'speech-ellipsis',
-  'switch-account',
-  'text-size',
-  'visit',
-];
-
 /**
  * [Icons](https://gestalt.pinterest.systems/web/icon) are the symbolic representation of an action or information, providing visual context and improving usability.
  *
@@ -95,7 +66,6 @@ const flipOnRtlIconNames = [
  * ![Icon dark mode](https://raw.githubusercontent.com/pinterest/gestalt/master/playwright/visual-test/Icon-list-dark.spec.ts-snapshots/Icon-list-dark-chromium-darwin.png)
  *
  */
-
 function Icon({
   accessibilityLabel,
   color = 'subtle',
@@ -104,38 +74,15 @@ function Icon({
   inline = false,
   size = 16,
 }: Props) {
-  const cs = classnames(
-    // @ts-expect-error - TS2345 - Argument of type 'string | undefined' is not assignable to parameter of type 'string'.
-    flipOnRtlIconNames.includes(icon) && styles.rtlSupport,
-    styles[color],
-    styles.icon,
-    { [styles.iconBlock]: !inline },
-  );
-  const isInVRExperiment = useInExperiment({
-    webExperimentName: 'web_gestalt_visualRefresh',
-    mwebExperimentName: 'web_gestalt_visualRefresh',
-  });
-  const path =
-    (icon && (isInVRExperiment ? vrIcons : icons)[icon]) ||
-    /* eslint-disable-next-line no-underscore-dangle */
-    (dangerouslySetSvgPath && dangerouslySetSvgPath.__path) ||
-    undefined;
-
-  const ariaHidden = accessibilityLabel === '' ? true : null;
-
   return (
-    // @ts-expect-error - TS2322 - Type '{ children: Element; "aria-hidden": true | null; "aria-label": string; className: string; height: string | number; role: "img"; viewBox: string; width: string | number; }' is not assignable to type 'SVGProps<SVGSVGElement>'.
-    <svg
-      aria-hidden={ariaHidden}
-      aria-label={accessibilityLabel}
-      className={cs}
-      height={size}
-      role="img"
-      viewBox="0 0 24 24"
-      width={size}
-    >
-      <path d={path} />
-    </svg>
+    <InternalIcon
+      accessibilityLabel={accessibilityLabel}
+      color={color}
+      dangerouslySetSvgPath={dangerouslySetSvgPath}
+      icon={icon}
+      inline={inline}
+      size={size}
+    />
   );
 }
